+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Inequality tolerance is now 1e-10 when validating solvers on README example
 - Refactor QP to SOCP conversion to use more than one SOCP solver
 - Rename "example problem" for testing to "README problem" (less ambiguous)
+- Rename `sw` parameter of `solve_safer_qp` to `sr` for "slack repulsion"
 - Reorganize code with a qpsolvers/solvers submodule
 - quadprog: warning when `initvals is not None` is now verbose